SmallCorrections

+++ escalade-model/src/main/java/com/aubrun/eric/projet6/model/bean/Photo.java
@@ -14,17 +14,17 @@ public class Photo {
     @Id
     @GeneratedValue( strategy = GenerationType.IDENTITY )
     @Column( name = "id" )
-    private Long   id;
+    private Integer id;
     @Column( name = "nom" )
-    private String nomPhoto;
+    private String  nomPhoto;
     @Column( name = "chemin" )
-    private String cheminPhoto;
+    private String  cheminPhoto;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId( Long id ) {
+    public void setId( Integer id ) {
         this.id = id;
     }
